reorganized rake tasks, added display name column

+++ db/migrate/20140726140330_create_projection_for_qbs.rb
@@ -40,8 +40,8 @@ class CreateProjectionForQbs < ActiveRecord::Migration
       t.integer :fumbles_low
 
       t.float :fantasy_points
-      t.integer :fantasy_points_high
-      t.integer :fantasy_points_low
+      t.float :fantasy_points_high
+      t.float :fantasy_points_low
 
       t.timestamps
     end
