+++ db/migrate/20140726161234_create_projection_for_tes.rb
@@ -20,8 +20,8 @@ class CreateProjectionForTes < ActiveRecord::Migration
       t.integer :fumbles_low
 
       t.float :fantasy_points
-      t.integer :fantasy_points_high
-      t.integer :fantasy_points_low
+      t.float :fantasy_points_high
+      t.float :fantasy_points_low
 
       t.timestamps
     end
